Modern programming & Study Hub
NoteVibe is a modern, web-based platform designed to streamline note-taking, study organization, and interview preparation. Built with Python, HTML, CSS, and Flask, it offers a clean, intuitive interface for users to  access notes, cheatsheets, and interview prep resources all in one place.

Key Features:
- Access curated cheatsheets for quick reference
- Get best Notes for quick revision
- Prepare for interviews with dedicated resources
- Responsive design for seamless use on any device
- User-friendly navigation and attractive UI animations

NoteVibe is ideal for students, professionals, and lifelong learners seeking a centralized, distraction-free environment for their study and productivity needs. The project is actively developed and open to contributions, with a focus on usability, reliability, and continuous improvement.